Update patch set 6
Patch Set 6: Code-Review-1

Agree the rule check should be added.
I'd also like a high level explanation of how VxLAN works with XenServer somewhere, since the tunnel endpoints aren't the same as with KVM.

I guess with devstack it might just work because the network node is running in the same place as the compute?

Patch-set: 6
Label: Code-Review=-1