Update patch set 14
Patch Set 14:

(3 comments)

> Like in sahara project we are widely using heat to provision infrastructure. It's clear that if your are using sahara in devstack, you have to enable additional plugin for heat always. 

This is the type of practical motivation that's nice to describe, so people can understand where things are actually going to be used.

> figuring out dependencies becomes hard.  

>> No, you just go through logs and it becomes clear what projects were fetched. It just adds logical encapsulation and makes stuff around enabling plugins more clear.

What I'm talking about is more like in project-config, where we enumerate the plugins required.  now you enable sahara, and get heat included, but that's not explicitly called out.  it makes it harder to follow where a plugin is actually used.  and I wonder about common things like selecting different branches or gitrefs in a situation like this, where you need to test different things for dependencies.

> i think i'd prefer to see it parsed
>> In such way it should be parsed? 

the plugin-deps file sort of looks like a config file, but it's being sourced and so eval'd as a bash file.  i think i'd prefer to see a cat/awk type split-up of the file so it's not evaluated as a bash file, and thus we avoid the possibility of other stuff creeping in there

i have a few other concerns inline

In general, I'm a bit wary that this might be an anti-pattern because automated dependency solving tends to hit a lot of corner cases (for example, the branching).  I may be over thinking it, I agree, but once this is in, we can never effectively unsupport it.

Patch-set: 14