WIP/DNM: Leave me alone, I'm concentrating...
Don't merge of course

Inform user that the logfile skips lines matching regex

Provide a notice at the top and bottom of the logfile informing the user
that the logfile does not print content that matches the specified
regular expression. At the bottom provide statistics on how many lines
were skipped based on the regex match.

Change-Id: I493014adf2a5a7325d5baa5f27592fa2239f1ab9