Update patch set 1
Patch Set 1:

Ian, true but none of that works yet for passing through local.conf via devstack-gate, and the long-term point of this change is to add a job to nova's experimental queue that runs with image signing and verification. I also think there are going to be more changes based on this conditional flag than just updating nova.conf.

Patch-set: 1
Label: Verified=0