Update patch set 1
Patch Set 1: Code-Review-1

I totally get why that's annoying ... but in the logs you now get a whole bunch of 

 2016-10-07 04:52:52.059 | +++ lib/apache:source:22                     :   set +o
 2016-10-07 04:52:52.059 | +++ lib/apache:source:22                     :   grep xtrace
 2016-10-07 04:52:52.062 | ++ lib/apache:source:22                     :   _XTRACE_LIB_APACHE='set -o xtrace'
 2016-10-07 04:52:52.063 | ++ lib/apache:source:23                     :   set +o xtrace

the whole point of this xtrace swizzling is that we're trying to find balance between usability/debugability and useless log spew.  So this re-introduces a bunch of boiler place noise back into the logs which hurts signal to noise ratio.

we could tighten the regex i suppose ... but is it worth it?  trying to find a balance between the noise in thousands of logs per day v the chance someone hits this

Patch-set: 1
Label: Code-Review=-1