Update patch set 1
Patch Set 1:

This is a reasonable change to avoid the current traceback that's happening, though we should address that as you've done in your oslo.db patch.

This also more closely matches how I think cells will be stood up in practice, setup the api and then add cells to it.

Patch-set: 1