Update patch set 1
Patch Set 1:

> I guess I am not a fan of "special cases". Why is xtrace so special? Why aren't other things ignored?

It's completely arbitrary, but practical in the sense that we try to balance unnecessary noise in the logs and information.  That's why we have a the xtrace on/off stuff at all -- it's hard enough to debug failures without (literally) hundreds/thousands of lines of boilerplate stuff to filter out.  this is filtering out very common boilerplate stuff that's in every file. 

> It just really bothers me that things can silently be thrown away from the logfile. As happened to me. Luckily I was able to figure out why after wandering from project to project. But others may not be so lucky.

I get it, but this just puts noise in the logs.  if you feel motivated enough to make it so that we don't put this noise in the logs, but still handle the case you faced, i think that's moving in the right direction.  adding useless stuff to the logs doesn't, IMO

Patch-set: 1