Update patch set 2
Patch Set 2:

To reiterate what the commit message has said. The whole point of this change is to explicitly use FIXED_RANGE and NOT use another variable because this is the source of the regression in the first place. The configuration API devstack previously exposes is to use FIXED_RANGE to avoid these overlaps. But the switch to completely independent variables ignores anyone who has previously set this variable to avoid conflicts.

I do not think we should do this because it is the source of the regression. We could indeed just edit the subnet pool variables and hope that we don't conflict anywhere or we can fix the problem by actually fixing it.

Patch-set: 2
Reviewer: Gerrit User 4146 <4146@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0