Update patch set 2
Patch Set 2:

> > > > I think this is not a regression, the error is introduced by
 > > [1],
 > > > > I'd almost be tempted to say that the actual fix is a revert
 > of
 > > > > that one.
 > > > >
 > > > > [1] https://review.openstack.org/#/c/356026/
 > > >
 > > > That would be one way to fix it, but aiui we already rely on
 > > larger
 > > > subnet pools for several tests (though I may have misunderstood
 > > > that) so there was a desire to roll forward and keep the tests
 > > > working rather than revert a whole bunch of things.
 > >
 > > The only tempest tests that need subnetpools are the
 > auto-topology
 > > ones and they only need a handful of subnets, not hundreds. To me
 > > the smoking gun [1] is telling me that we merged it without fully
 > > understanding the consequences. Besides, I am not sure why that
 > > patch was allowed in the first place. Forcing subnet creations
 > out
 > > of a subnet pool when available is clearly altering how devstack
 > > always worked and I would have been opposed to it without fully
 > > understanding the actual need behind it.
 > 
 > You should be able to propose the revert of [1] and have it be self
 > tested. Just have to recheck if the multinode jobs don't land in
 > osic.

Also we really really need documentation for this stuff especially if it is going to be turned on by default. It isn't clear at first glance if neutron is setting the route or devstack (looks like devstack based on the proposed possible revert) or why editing the routing table is desired in the first place (allows VMs to talk to hypervisors via their tenant networks which is weird).

If we decide that we cannot re consume FIXED_RANGE as the designator for a safe subnet range here then we also need to make sure we document how to avoid these conflicts (what vars to set) and make sure we set them for the gate jobs before we start relying on this. And probably have a big warning for "Hey this changed you may need to set these new vars" in release notes or something.

Patch-set: 2