Update patch set 4
Patch Set 4:

@Ian,

Doing some measurements and I was surprised that the regex eliminates about 17% of the logfile line count! That seems like a lot to me.

So I agree the skipping is a good thing. I am proposing adding a notice to the logfile to inform the user about this skipping.

Thanks for your help.

Patch-set: 4
Label: Verified=0