Update patch set 2
Patch Set 2:

The reason behind review.openstack.org/356026 was that the v4 subnet pool was being overlapped by the FIXED_RANGE, and all the routing etc was based on the FIXED_RANGE. Reverting the change isn't helpful in that sense.

We wanted to make use of subnet pools, and my belief was that they should be used by default (that was a mistake). Why would we encourage the manual management of subnets when we have a mechanism that does the work for us?

The patch missed two points though - keeping the current behaviour as default and taking a closer look at the subnet pool range. Between this proposed patch, once an agreement has landed about the values to use, and review.openstack.org/378063, these points will be settled.

Patch-set: 2
Reviewer: Gerrit User 15321 <15321@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0