Stop using nova-manage network create
nova-manage network create is deprecated, so this replaces
it with the equivalent nova CLI from python-novaclient.

Because the nova CLI hits nova-api and nova-api calls
nova-net to create the network, which calls conductor
to use the database, we have to start nova-conductor and
nova-net before we can create the network now.

Change-Id: I5c4291509841325e6123520131f30b64c847a17f