Update patch set 1
Patch Set 1: Code-Review-1

I put a comment in the linked bug, but this appears to come from the ironic plugin, not devstack.  you can add it to the list of packages there ... but it also seems the command is deprecated for "ip link" so probably you can change the ironic plugin to not need it at all.

Might be nice to remove it in devstack, too, if that's the only reason it's there ... can't see any neutron or nova-net references for example.

Patch-set: 1
Reviewer: Gerrit User 7118 <7118@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1