Stop using nova-manage network commands
The nova-manage network-related commands are deprecated,
so this replaces them with equivalent nova CLI from
python-novaclient.

Because the nova CLI calls nova-api and nova-api calls
nova-net to create the network, which calls conductor
to use the database, we have to start nova-conductor and
nova-net before we can create the network now.

This adds a new idempotent 'start_nova_process' function
which is used to start n-cond and n-net after n-api but
before creating the network and floating IP resources.
The start_nova_rest function is changed to use start_nova_process
because grenade calls start_nova which calls start_nova_rest.

In the case of full grenade, n-cond and n-net won't be running
on the new (ocata) side, so start_nova will start them.

In the case of partial grenade and non-grenade jobs, however,
n-cond and n-net will already be running when start_nova is
called.

Change-Id: I5c4291509841325e6123520131f30b64c847a17f