Revert "Enable neutron to work in a multi node setup" partly
This reverts a part of commit 88f8558d874072536e7660a233f24207a7089651
because it was wrong.  A better fix will follow.

The configuration in question uses the new lib/neutron. (not neutron-legacy)
In that case, calling create_neutron_initial_network from stack.sh directly
is a wrong way, as create_neutron_initial_network is sourced by
neutron-legacy.  The new neutron code should not rely on the legacy one.

Change-Id: Ib73f5a88c5a00d6748e10641a6b5abed16967cc3