Add support for containerized ceph
This commit is a follow up of https://review.openstack.org/#/c/374825/
which introduced the support to containerized Ceph.
If CEPH_CONTAINERIZED=True, to apply the approprieate changes
into Ceph we need to access the Docker container.
Usually DOCKER_EXEC variable contains: "docker exec ceph-demo" which
then connect inside the container and run the desired command. If the
variable is empty this won't break existing configuration.

Change-Id: I24a1fc4fb65cf9770b664539db013c8311c2ca5f
Signed-off-by: Sébastien Han <seb@redhat.com>