Update patch set 5
Patch Set 5: Code-Review-1

overall this seems like a good idea to get rid of some of the larger unneeded images.

however, I don't really agree with the deprecation part.  firstly nobody reads lines like that mixed in thousands of other lines of output [1], but secondly we should default to "just work" as much as possible.

So I'd support making the default able to be overridden for people who know what they in cases like infra, with added documentation about how to use this so people can actually discover how to trim the list (I guess it fits in configuration.rst) 

[1] we actually have a "deprecated" function to call this stuff out at least into the warnings summary file, but not convinced we actually have a clear way to deprecate this anyway.

Patch-set: 5
Label: Code-Review=-1