Update patch set 1
Patch Set 1:

> This doesn't reset the global value (the one in the DEAFULT section) because this value was and is not taken into account at all. 

I don't see that this is set specifically in the LVM volume_opts [1], and it seems to just get the global value (_delete_volume).  I6afbdecea637f7e7d3417be537ab71fe97bffda3 suggests that the value in volume_ops is set from the global value?

if it is the case that setting it globally doesn't actually get picked up, i feel that is certainly not made clear in any documentation i could see

[1] http://git.openstack.org/cgit/openstack/cinder/tree/cinder/volume/drivers/lvm.py#n46
[2] http://git.openstack.org/cgit/openstack/cinder/tree/cinder/volume/drivers/lvm.py#n131

Patch-set: 1