Update patch set 13
Patch Set 13: Code-Review+2

I made minor changes to the configuration.rst for wording and updated the changelog to reflect what's happening in the latest change.

There's been a couple of things coming up that are around plugins for nova drivers which to my mind aren't fully defined.  If we wanted plugins getting their images into this canonical list I think there's more work to.  But IMO that should be based on this basic principle, so LGTM

Patch-set: 13
Label: Code-Review=+2