Ability to force pre-cache of test images
Add DEVSTACK_NO_DOWNLOAD_IMAGE to prevent automatic downloads of test
images.  This is for situations where images should be pre-cached and
we wish to ensure we are not hitting the network to download during
test runs.

Note when in use with devstack-gate, this is a bit of a chicken-egg
problem for adding new downloads (gate would have to cache them before
they are usable).  Initial testing can easily override this
explicitly.  At commit time, we would need to commit the image update
first, allow a new build of the infra-images and then commit whatever
is using the image (or, just accept the new test is broken until the
new daily image builds with the download in it).  Given the
infrequency at which this occurs and worst-case latency of 24 hours
(next image build), I don't see this being a big issue.

Change-Id: Iace57fd0042c35535e72a3b316ea594516968113