Update patch set 11
Abandoned

I think we basically can't do this right now because with cells v1 the network and floating IP pool has to be created in the child cell directly, the API doesn't know where to route the network create request. So nova-net is around as long as cells v1 is around, so the nova-manage network create command needs to be also.

Patch-set: 11
Status: abandoned