Update patch set 10
Patch Set 10:

> This changes OVS_PHYSICAL_BRIDGE to PUBLIC_INTERFACE yet fails to
 > update any of our documentation.

updated commit message:

L3 correct bridge for external from OVS_PHYSICAL_BRIDGE to PUBLIC_BRIDGE

by default "OVS_PHYSICAL_BRIDGE" and "PUBLIC_BRIDGE" are set to "br-ex"
that mean we use 1 bridge in OpenStack.

This patch fix l3 to allow a case when external and internal networking
are different bridges and different interfaces.

The patch also retain the old behavior because both parameters are set to
same bridge name by default.


----

In mind "l3" seem to be designed to be with this patch but someone probably forgot why he defined "PUBLIC_BRIDGE".
This patch will allow to defined different bridges for internal and external networking while retain same flow when it is the same bridge.

I follow the guides and they seem updated and not need to update any current documents.

Patch-set: 10
Label: Verified=0