Skip create flavors when only n-api is enabled.
Zero Downtime upgrade process enables n-api on subnode along with n-cpu.
When n-api is enabled on subnode, devstack tries to create flavors
resulting in errors as keystone is not available on subnode. To avoid
this issue, check is added to see if keystone is enabled along with
n-api on subnode before creating flavors.

Change-Id: I1adeb20f261004a24d3f28723ccbf84699c53a19