Add firefox v46 deb package to the image list
Horizon project is no longer able to use the uptodate firefox version
that came with xenial/trusty (49 as the time of committing this
change), because they broke compatibility with selenium firefox
webdriver in version 47, see
https://github.com/SeleniumHQ/selenium/issues/2110

As a solution we donwloaded previous firefox version 46 from
https://sourceforge.net/projects/ubuntuzilla/files/mozilla repo. The
downside of that were ftp connection problems which lead to tests
instability. The aim of this change is to avoid connection problems
with sourceforge ftp and use older version of firefox in selenium
tests at the same time.

Change-Id: I1521198146aa0ce332a59e8b40c5afe9541b7c03