Update patch set 1
Patch Set 1:

(follow up to a short conversation on IRC)

Is UUID-formatted token tested anywhere (not unit tests).  If that isn't deprecated then it should be somewhere.  And I don't think we can merge this if we need to test UUID tokens due to the test to skip the fernet initialization.  If running that all the time is harmless then maybe?

Patch-set: 1
Reviewer: Gerrit User 970 <970@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0