nova: add support for TLS between novnc proxy & compute nodes
Nova is gaining the ability to run TLS over the connection
between the novnc proxy service and the QEMU/KVM compute node
VNC server.

This adds a new config param

NOVA_CONSOLE_PROXY_COMPUTE_TLS=True

which instructs devstack to configure libvirt/QEMU to enable
TLS for the VNC server, and to configure the novncproxy to
use TLS when connecting.  NB this use of TLS is distinct from
use of TLS for the public facing API controlled by USE_SSL,
they can be enabled independently.

This is done in a generic manner so that it is easy to extend
to cover use of TLS with the SPICE and serial console proxy
services too.

Implements bp: websocket-proxy-to-host-security
Depends-on: I9cc9a380500715e60bd05aa5c29ee46bc6f8d6c2

Change-Id: Ib29d3f5f18533115b9c51e27b373e92fc0a28d1a