WIP/DNM: Testing
This is a temporary patch until issue with multiple enables is fixed.

If plugin is enabled multiple times only register 1st time

When using the enable_plugin command and grenade jobs it can be
easy to enable the same plugin twice, as the grenade job has a
registration section and the configuration in project-config can also
enable it due to code-reuse in project-config.

This change makes it so if it sees the same plugin name is enabled
more than once only the first entry will be enabled and registered.

Change-Id: Iad3301e3ec7e4365f56e674d86fccb89cc134d5b