If plugin is enabled multiple times only register 1st time
When using the enable_plugin command and grenade jobs it can be
easy to enable the same plugin twice, as the grenade job has a
registration section and the configuration in project-config can also
enable it due to code-reuse in project-config.

This change makes it so if it sees the same plugin name is enabled
more than once only the first entry will be enabled and registered.

Change-Id: I9f1d7e58b861b04473b6a57c9ad404203fb7277a
(cherry picked from commit 923626a25bc49398d80ace7654584525399fb94e)