WIP/DNM: Testing
This is a temporary patch until issue with multiple enables is fixed.

If plugin is enabled multiple times only register 1st time

When using the enable_plugin command and grenade jobs it can be
easy to enable the same plugin twice, as the grenade job has a
registration section and the configuration in project-config can also
enable it due to code-reuse in project-config.

This change makes it so if it sees the same plugin name is enabled
more than once only the first entry will be enabled and registered.

Change-Id: I9a72b287e15d381e261b1e03ea9911f50a56ff6f
(cherry picked from commit 0b492fe4b6c716b668455bca96641e98c07a94f1)