Update patch set 1
Patch Set 1: Code-Review-1

this is supposed to be a generic tool, not to have undocumented workarounds like this put in it.

the idea seems ok if we want this pre-cached, but it needs to be in stack.sh along with a comment.

what's the long term plan?

Patch-set: 1
Reviewer: Gerrit User 7118 <7118@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1