Update patch set 3
Patch Set 3:

(1 comment)

i guess the other choice is to abort when we see this with a clear message.

I'm really not sure if plugin enablement is supposed to be idempotent?  if it is, then ignoring multiple enables is probably right.  if it isn't, then aborting is probably right.

Patch-set: 3
Label: Verified=0