Update patch set 3
Patch Set 3:

(1 comment)

I can't really speak to if it is supposed to be idempotent or not. I would like it to be as I am running into this issue with our ironic multitenant grenade experimental job.

But accepting multiple entries does cause the job to fail in ways that are not immediately obvious. As in it took me awhile to realize what the problem was :(

Patch-set: 3
Reviewer: Gerrit User 14760 <14760@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0