Update patch set 3
Patch Set 3: Code-Review-1

unfortunately I think something is going on with ssl certs here.  the suspicious warnings are in [1]

the updown test is a bit silly, in that it's testing cleanup so it goes ahead and deletes all the log files.  ergo it's hard to compare a failed instance to a working instance, because the working instance has no logs.

but, i think there's something going on with the ca bundle setup, maybe because the calls have moved.  in [2] you've got

  2016-11-14 01:14:09.998 | + lib/tls:fix_system_ca_bundle_path:368    :   capath=/etc/ssl/certs/ca-certificates.crt

and in [3] for comparison it is 

  2016-11-14 01:04:07.173 | + lib/tls:fix_system_ca_bundle_path:349    :   capath=/usr/local/lib/python2.7/dist-packages/requests/cacert.pem

therefore, i think it might have something to do with ordering and pip install of pyopenssl or something?

why does this only affect the "updown"?  maybe because it's the only one leaving TLS defaults alone?

but yeah, i think there's a real problem lurking


[1] http://logs.openstack.org/86/396186/3/check/gate-devstack-dsvm-updown/596161c/logs/screen-g-api.txt.gz#_2016-11-14_01_29_00_120
[2] http://logs.openstack.org/86/396186/3/check/gate-devstack-dsvm-updown/596161c/logs/devstacklog.txt.gz#_2016-11-14_01_14_09_998
[3] http://logs.openstack.org/44/396644/1/check/gate-devstack-dsvm-updown/a594338/logs/devstack-early.txt.gz#_2016-11-14_01_04_07_173

Patch-set: 3
Label: Code-Review=-1