Update patch set 3
Patch Set 3:

(1 comment)

I'd agree multiple times appearing in the list is wrong.

My tendency would to be pretty conservative in situations like this, and error out if multiple things enable the same plugin -- because they might be doing it with different and conflicting ideas of what they're trying to set up.  It seems like something you'd have to sort out.  But I can see other arguments

I'm just cautious because once we move from this undefined state to codify one way or the other, it's generally hard to go back once people start relying on that

Patch-set: 3