remove listen directive bound to only ipv4
Using devstack on a  RHEL based system results in "Listen 0.0.0.0:80" being added to the /etc/httpd/conf/httpd.conf. This configures Apache to only listen to port 80 on an IPv4 interface. This makes it not possible to access Horizon via IPv6 without re-configuring and restarting httpd.

Removing this sed leaves the default "Listen 80" from the rpm package, which binds to all interfaces and will allow connection to Horizon via IPv6.

Change-Id: I9fe8cbebff0ca6a30ceeaae0f7e035c9bb828d44