Update patch set 2
Patch Set 2: Code-Review+2

(1 comment)

After a bit of digging, this seems sensible to me.  I say that because it's probably a better solution to bug #1628267 than the one we merged.

I'd appreciate it if the documentation got some help as mentioned inline.  I think the idea when subnetpools were added was that they were essentially transparent, but it seems that's not quite the case.  Basically, AIUI, it needs to be made clear that FIXED_RANGE is a deprecated nova-net thing, and that for neutron the safe-to-use addrs will be put into the subnetpool and networks allocated from there (by default).

I also wonder if this could provide us reverting Ic89ceca76afda67da5545111972c3348011f294f ... to actually make it transparent again.  I think it is quite reasonable to say "we use subnet pools" to simplify all this rule and just work out the issues so everyone can use it 

I hope to see that as clarifying follow-ons, but I think this is probably headed in the right direction.

Patch-set: 2
Reviewer: Gerrit User 7118 <7118@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2