Update patch set 2
Patch Set 2:

(1 comment)

Yeah, I think having the discussion about eliminating non-subnetpool automatic private networks in devstack is reasonable. If someone wants a private network not using the default subnetpool they could always add one.

Really the thing that's screwed up is the automatic routing we are doing for private v4 networks in devstack. This isn't how a normal neutron deployment works and I'd like to eliminate it here: https://review.openstack.org/#/c/397987/

Once that is gone, we can care a whole lot less about what is used in a private network because it will no longer conflict with the host.

Patch-set: 2