Change FIXED_RANGE references to FIXED_RANGE_V6
In the 'Private Network Addressing' section of the doc,
there are references to FIXED_RANGE when referring to V6
networks.  These have been changed to FIXED_RANGE_V6.

Also fixed a few typos and grammatical errors when
giving the doc a quick read-through looking for more
references to FIXED_RANGE.

Change-Id: Iaa530c476ce2b36a3f616945ddd2e24fa599a16c