Apache fallback from reload
Since 4b49e409f853104dae021dfca1a9342ec9ac4709 devstack
started to use reload instead of restart.

Using reload in devstack,
does not seams to be the wisest idea,
because we do not need greacful restart here.

Systemd also reject to reload something,
which was not loaded before.

 $ sudo /bin/systemctl reload httpd
 httpd.service is not active, cannot reload.

This change just adds a fallback to restart,
in case the reload got rejected for whatever reason.

Change-Id: I70d597fbe4a8923d937ba8432e29edefb27d1058