Update patch set 12
Patch Set 12:

> (1 comment)

we are not setting "lockout_duration" to 1800 since would not be feasible to test it with such high number. That's why we are using 5.

"unique_last_password_count" has the default set to 1, which means it is disabled - the setting only makes sense to be enforced if using something greater than 1.

Patch-set: 12