Update patch set 2
Patch Set 2:

Those are the exact steps we need to write tests for. Storing a cert in the key manager requires a test, uploading a signed image to Glance is a test, booting a signed image is another test. If these operations are handled by devstack and fail, we won't know the point of failure.

Patch-set: 2