Update patch set 2
Patch Set 2:

Dane, as I was confused about what Sean was asking for, I did some digging and I'm pretty sure it's this:

http://docs.openstack.org/developer/devstack/configuration.html#local-conf

Which I think means, rather than add a variable to devstack that sets nova.conf values, we could do this in the local.conf which gets setup by the job. There were some changes Sean was working on to pass post-config values down through devstack-gate to devstack, but I'm not sure if that's all done now.

Patch-set: 2