Set the reseller config of Tempest disabled
On Keystone side, the extension will be always enabled in Ocata by
removing the extension config. On the other hand, the corresponding
Tempest config was disabled as the default value for old branches
(Liberty and Kilo at the time).
However, the Tempest default value should be changed to enabled
for matching to Keystone development. So we need to set the config
as disabled clearly on Liberty branch even but the EOL is close.

Change-Id: I7ee81396367d2c6259cb6d55b8ce83bec92c6562
Needed-By: Id3dd322b3a0585ed95eb2dea4ad35a7949bb9b1f