Update patch set 1
Patch Set 1:

Hello, @Sean Dague. Thank you for merging another fix [1] which is almost doing the same thing as this fix.
Because the goals between [1] and this fix are the same, so I would like to abandon this fix.
However, I don't think you are doing the right thing.
If there is an existing open patch, we should never start a new one for the same topic.
I know we are at the opposite side on earth, and this caused my response late, but we should stick to the community way to do things here, I think.
[1] https://review.openstack.org/#/c/405446/

Patch-set: 1