Update patch set 5
Patch Set 5:

The issue here is https://bugs.launchpad.net/ubuntu/+source/linux/+bug/759725, which is marked as won't fix. Basically ubuntu wants to keep the kernel readable by root only, and libguestfs thinks that's not right. The only proposed workaround it to relax read restrictions on the kernel, which is what this patch is about.

Patch-set: 5