Update patch set 2
Patch Set 2: Code-Review-1

This seems like a weirdly abstracted way to do this.  I am not multi-node-multi-keystone-rolling-upgrade-grenade expert but the solution here seems to be to *not* call recreate_db when you know it shouldn't be called.  surely at the the point we're calling this, we can check if it should be called and just skip it?

Patch-set: 2
Reviewer: Gerrit User 7118 <7118@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1