stack.sh doesn't support shared nova instance path
Nova's instances directory may be a shared directory (ie nfs), in
these cases, we do not want to clean out the directory.

This patch introduces the parameter SHARED_INSTANCE_DIR. This value can
be set to True or False (defaults to False). If set to True, nova instances
directory is treated as shared and will not be deleted; otherwise, it will be
deleted when stack.sh is invoked.

Change-Id: I644fc45072e94f53dfe548c92a67ee05f08ba7d9
Closes-Bug: #1649389