stack.sh doesn't support shared nova instance path
In cases where the nova instances directory is an (nfs) shared directory
and nova instances path is located somewhere under nova state path,
running stack.sh for a multinode setup and for a compute node results
in an error message when stack.sh tries to cleanup.

This fix unmounts and remounts the nfs directory as needed during the
cleanup.

Change-Id: I644fc45072e94f53dfe548c92a67ee05f08ba7d9
Closes-Bug: #1649389