Update patch set 4
Patch Set 4: Code-Review-1

(1 comment)

See also within for why I think this solution is less than ideal.

The reason there was ever a VirtualHost for placement was because of trying to have logs that are separate from the logs of other services hosted by httpd. If we are willing to get rid of that constraint (and simply log apache2 to its own normal logs) we can effectively host WSGI apps under prefixes without much complexity in the template files used for each service. Log separation is easy enough to do in post-processing (by the power of grep and friends) if we need to?

Patch-set: 4
Label: Code-Review=-1