Update patch set 1
Patch Set 1: Code-Review-1

So I understand the issue now, but I think this solution might be a bit too specific for your use case. I'm thinking that instead of implicitly checking for nfs under the covers if it would just be better to have a config variable like SHARED_INSTANCE_DIR (maybe not the best name) that will trigger this behavior when true. Well, not the unmounting and remounting, but the more selective deleting on teardown.

My thinking is that there are other types of shared file systems that could be used for the instance dir and they'll all have this same issue, but this patch only works if it's nfs.

Patch-set: 1
Label: Code-Review=-1