Update patch set 12
Patch Set 12:

fwiw, that paste is incomplete, the changes made to the first script need to mirrored for all of them.

But doing that will maintain all the current functionality that exists today, but when swift is installed the bin/* scripts will have a shebang for the version of python used to do the install.

Patch-set: 12