allow config to manage python3 use explicitly
Add variables ENABLED_PYTHON3 and DISABLED_PYTHON3 to work like
ENABLED_SERVICES and DISABLED_SERVICES and to manage which packages are
installed using Python 3. Move the list of whitelisted packages in
pip_install to the default for ENABLED_PYTHON3, except swift which is
not enabled by default for now.

Add enable_python3 and disable_python3 functions to make editing the
variables from local.conf easier.

Add python3_enabled_for and python3_disabled_for functions to check the
settings against packages being installed by pip.

Update pip_install to check if python3 is disabled for a service, then
see if it is explicitly enabled, and only then fall back to looking at
the classifiers in the packaging metadata.

Update pip_install messages to give more detail about why the choice
between python 2 and 3 is being made for a given package.

Change-Id: I69857d4e11f4767928614a3b637c894bcd03491f
Signed-off-by: Doug Hellmann <doug@doughellmann.com>