Update patch set 12
Patch Set 12:

> I still don't like loosening these checks and adding this new
 > behavior.

These check are for creating a the mapping stay safe. I only allow to create/make sure "OVS_PHYSICAL_BRIDGE" if user has a custom mapping set in "OVS_BRIDGE_MAPPINGS". agree with me that if we got to these line we need to have the OVS_PHYSICAL_BRIDGE created or to make sure it is this is what the "_neutron_ovs_base_add_bridge $OVS_PHYSICAL_BRIDGE" do.

Patch-set: 12
Label: Verified=0