Update patch set 2
Patch Set 2:

This change is needed in stable/newton as grenade upgrades from newton to master in multinode gate jobs. This change avoids credential issues with creating flavors when only n-api is enabled on subnode.

Zero Downtime upgrade process enables n-api on subnode along with n-cpu.
When n-api is enabled on subnode, devstack tries to create flavors
resulting no credentials error as keystone is not available on subnode.

Patch-set: 2