Fix cellsv2 cell0 database name
For some reason we were defaulting the name of the cell0 database
to nova_api_cell0 instead of nova_cell0. Devstack inherited that to
make things work, but we don't really want that. This patch makes us
use the proper name and create the cell0 mapping accordingly. As a
side effect, it also starts the process of unifying the cellsv1 and
cellsv2 paths by creating the cell0 mapping the same for both.

Conflicts:
        lib/nova

NOTE(mriedem): The conflicts are due to not supporting cells v2
in a cells v1 deployment in Newton and the fact that cells v2
was still optional in Newton (but required in Ocata).

Change-Id: I4e7f6c5eaa068c98e5c4ef3feaee50d8e4f5d484
(cherry picked from commit 29bb53fd3e7762c1fea023842ba09fcbfab60252)