Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

I thought the intention of the map_cell0 default was to start people off with cell0 database co-located with the API database, as both are API-cell-level constructs. One could expect them to be on the same controller host, with the same user/password, as a default.

Patch-set: 1