Revert "lib/tempest: allow tweaking volume-feature-enabled/manage_snapshot"
As mentioned several other places, this is not something
that should be enabled by default. The input to this
functionality is backend specific. The issue is not that
"only a few Cinder backends support" it, but that the test
is flawed in assuming everything doing manage/unmanage
will take the same values to identify the volume or
snapshot to be managed.

This is causing many third party CI failures and should
be reverted back to defaulting to false.

This reverts commit afac732d5d9b03da7abbd61f814c2f0d1a9039da.

Change-Id: Id0fb170294d127f923b4bde487eed0e0755a0f71