Update patch set 1
Patch Set 1:

Jordan, I think the issue is more that the test itself is flawed and maybe we should go back to the patch I put up to revert it completely.

As is, this is only a test for the LVM driver. So it should either be updated to only run if the backend is LVM, or it should be removed and we should just stick to a functional test in cinder.

There are actual deployments that use LVM, but in probably the majority of the cases, if someone were to run this to validate their cloud, they are going to get incorrect and confusing results.

Granted, if using devstack it is easy to set a flag, but this is just yet another non-obvious detail that most will not know until they hit a failure and waste time (both theirs and others by needing to ask around to find out the answer) to know they need to change the default config to add a flag.

It does not make sense to me to have a default that works for one at the expense of many, rather than having something that is good for many with just one needing to change.

Patch-set: 1
Reviewer: Gerrit User 11904 <11904@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0